+++ src/main/java/racingcar/controller/RacingGameController.java
@@ -1,7 +1,5 @@
 package racingcar.controller;
 
-import java.util.List;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -16,6 +14,8 @@ import racingcar.dto.PlayRequestDto;
 import racingcar.dto.PlayResultResponseDto;
 import racingcar.utils.NumberGenerator;
 import racingcar.utils.RandomNumberGenerator;
+
+import java.util.List;
 
 @RestController
 public class RacingGameController {
